Attempt to fix subtle bug in player object, changed times to isoformat
When using 'until' field when looking up player objects there was a condition where API would report that the player had been on for day because the player logout would be cutoff and the API would extrapolate this time mean the player never logged out.

Attempting to fix this behavior by grabbing all events related to a player, parsing them normally, and _then_ cull by 'until'.